requirements.txt
requests
beautifulsoup4
icalendar
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
markdownify
openai
chromadb
tiktoken
fastapi
uvicorn
jinja2
python-multipart
pydantic
GitPython
schedule
anthropic
